+++ app.js
@@ -1,20 +1,19 @@
 const express = require("express");
-const app = express();
-const port = 3000;
-const router = require("./routes");
 const passport = require("passport");
 const cookieParser = require("cookie-parser");
+const router = require("./server/routes/routes");
+const { passportConfiguration } = require("./config/passportConfiguration");
 
-const { passportConfiguration } = require("./passportConfiguration");
+const app = express();
+const port = 3000;
 
 passportConfiguration();
 
 app.use(passport.initialize());
-
 app.use(cookieParser());
 app.use(express.urlencoded({ extended: true }));
-
-app.use(express.static("public"));
+app.use(express.static(__dirname + "/app"));
+app.use(express.static("app/homepage"));
 
 app.use(router);
