+++ app/homepage/calculateSol.js
