+++ app/homepage/displayPhoto.js
