+++ app/homepage/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Mars Rover</title>
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="/styles/style.css" />
     <script type="text/javascript" src="displayPhoto.js"></script>
     <script type="text/javascript" src="displayUsername.js"></script>
   </head>
@@ -27,11 +27,10 @@
       <p id="photo-explanation"></p>
     </section>
     <section class="link-container">
-      <a class="login-link" href="./login.html">
-        Login to your Mars Rover account
-      </a>
+      <a class="login-link" href="/login">Login to your Mars Rover account</a>
+      <a href="/rover">See the Mars Rover Map</a>
     </section>
-    <script type="text/javascript" src="script.js"></script>
+    <script type="text/javascript" src="displayTimes.js"></script>
     <script type="text/javascript" src="calculateSol.js"></script>
   </body>
 </html>
