+++ app/login/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Mars Rover</title>
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="../styles/style.css" />
   </head>
   <body>
     <h1>Login</h1>
@@ -27,7 +27,7 @@
         required
       />
       <button type="submit" value="submit">Login</button>
-      <a href="register.html">Register an account</a>
+      <a href="/register">Register an account</a>
     </form>
   </body>
 </html>
