+++ app/register/checkPassword.js
