+++ app/rover-map/mars-rover-sprite.png
