+++ app/rover-map/path-of-mars-rover-mobile.png
