+++ server/middleware/checkUserCredentials.js
